+++ config/initializers/config.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+#  Copyright (c) 2025, Schweizer Wanderwege. This file is part of
+#  hitobito and licensed under the Affero General Public License version 3
+#  or later. See the COPYING file at the top-level directory or at
+#  https://github.com/hitobito/hitobito_sww
+
+Rails.application.config.after_initialize do
+  require_relative "../../config/settings_contract_sww"
+  Config.validation_contract = SettingsContractSww.new
+  Settings.reload!
+end